requirements.txt
anyio==4.12.0
av==16.0.1
certifi==2025.11.12
click==8.3.1
coloredlogs==15.0.1
coverage==7.13.0
ctranslate2==4.6.2
faster-whisper==1.2.1
filelock==3.20.1
flatbuffers==25.12.19
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
humanfriendly==10.0
idna==3.11
iniconfig==2.3.0
mpmath==1.3.0
numpy==2.4.0
onnxruntime==1.23.2
packaging==25.0
pluggy==1.6.0
protobuf==6.33.2
Pygments==2.19.2
pyrefly==0.46.1
pytest==9.0.2
pytest-cov==7.0.0
pytest-mock==3.15.1
PyYAML==6.0.3
ruff==0.14.10
setuptools==80.9.0
shellingham==1.5.4
sympy==1.14.0
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.21.0
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
